+++ test/unit/src/lib/deleteHandlerTests.js
@@ -395,4 +395,22 @@ describe(('deleteHandler'), function testDeleteHandler() {
                 assert.notStrictEqual(deletedPaths.indexOf('/tm/net/tunnels/tunnel/~Common~tunnel'), -1);
             });
     });
+
+    it('should skip dns resolver f5-aws-dns on attempt to delete it', () => {
+        const declaration = {
+            Common: {
+                DNS_Resolver: {
+                    'f5-aws-dns': {},
+                    resolver: {}
+                }
+            }
+        };
+
+        const deleteHandler = new DeleteHandler(declaration, bigIpMock);
+        return deleteHandler.process()
+            .then(() => {
+                assert.strictEqual(deletedPaths.indexOf('/tm/net/dns-resolver/~Common~f5-aws-dns'), -1);
+                assert.notStrictEqual(deletedPaths.indexOf('/tm/net/dns-resolver/~Common~resolver'), -1);
+            });
+    });
 });